SE-08 COMPLETE

+++ src/main/java/ru/iokhin/tm/command/task/TaskRemoveAllByProjectCommand.java
@@ -2,7 +2,6 @@ package ru.iokhin.tm.command.task;
 
 import lombok.NoArgsConstructor;
 import org.jetbrains.annotations.NotNull;
-import ru.iokhin.tm.Bootstrap;
 import ru.iokhin.tm.command.AbstractCommand;
 
 @NoArgsConstructor
@@ -12,14 +11,6 @@ public final class TaskRemoveAllByProjectCommand extends AbstractCommand {
     public boolean security() {
         return true;
     }
-
-//    private boolean isHaveAccess(Bootstrap bootstrap, String projectId) {
-//
-//        @NotNull final String currentUserId = bootstrap.getCurrentUser().getId();
-//        @NotNull final String allowedUserId = bootstrap.getProjectService().findOne(projectId).getUserId();
-//
-//        return currentUserId.equals(allowedUserId);
-//    }
 
     @Override
     public String name() {
@@ -33,17 +24,13 @@ public final class TaskRemoveAllByProjectCommand extends AbstractCommand {
 
     @Override
     public void execute() {
+        bootstrap.getCommandMap().get("project-list").execute();
         System.out.println("ENTER ID OF PROJECT TO CLEAR TASKS");
-        @NotNull String projectId = bootstrap.getTerminalService().nextLine();
-//        if (bootstrap.getProjectService().findOne(projectId) == null) {
-//            System.out.println("NO SUCH PROJECT ID");
-//            return;
-//        }
-//
-//        if (!isHaveAccess(bootstrap, projectId)) {
-//            System.out.println("NO ACCESS FOR THIS OPERATION");
-//            return;
-//        }
+        @NotNull final String projectId = bootstrap.getTerminalService().nextLine();
+        if (!bootstrap.getTaskService().removeAllByProjectId(bootstrap.getCurrentUser(), projectId)) {
+            System.out.println("NO SUCH PROJECT ID");
+            return;
+        }
         bootstrap.getTaskService().removeAllByProjectId(bootstrap.getCurrentUser(), projectId);
         System.out.println("OK");
     }
