+++ src/main/java/ru/iokhin/tm/repository/TaskRepository.java
@@ -1,5 +1,7 @@
 package ru.iokhin.tm.repository;
 
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import ru.iokhin.tm.api.repository.ITaskRepository;
 import ru.iokhin.tm.entity.Task;
 
@@ -9,9 +11,9 @@ import java.util.Collection;
 public class TaskRepository extends AbstractRepository<Task> implements ITaskRepository {
 
     @Override
-    public Collection<Task> findAllByUserId(String userId) {
-        Collection<Task> taskCollection = new ArrayList<>(0);
-        for (Task task : findAll()) {
+    public Collection<Task> findAllByUserId(@NotNull final String userId) {
+        @NotNull final Collection<Task> taskCollection = new ArrayList<>(0);
+        for (@NotNull Task task : findAll()) {
             if (task.getParentId().equals(userId))
                 taskCollection.add(task);
         }
@@ -19,21 +21,23 @@ public class TaskRepository extends AbstractRepository<Task> implements ITaskRep
     }
 
     @Override
-    public void removeAllByUserId(String userId) {
+    public void removeAllByUserId(@NotNull final String userId) {
         for (Task task : findAllByUserId(userId)) {
             remove(task.getId());
         }
     }
 
     @Override
-    public Task findOne(String parentId, String id) {
-        Task project = repository.get(id);
-        return project.getParentId().equals(parentId) ? project : null;
+    public Task findOne(@NotNull final String parentId, @NotNull final String id) {
+        @Nullable final Task task = repository.get(id);
+        if (task == null) return null;
+        return task.getParentId().equals(parentId) ? task : null;
     }
 
     @Override
-    public Task remove(String parentId, String id) {
-        Task task = repository.get(id);
+    public Task remove(@NotNull final String parentId, @NotNull final String id) {
+        @Nullable final Task task = repository.get(id);
+        if (task == null) return null;
         return task.getParentId().equals(parentId) ? repository.remove(task.getId()) : null;
     }
